common: avoid signed-shift overflow in MSG_ReadLong

+++ src/common/movemsg.c
@@ -1210,10 +1210,10 @@ MSG_ReadLong(sizebuf_t *msg_read)
 
 	else
 	{
-		c = msg_read->data[msg_read->readcount]
-			+ (msg_read->data[msg_read->readcount + 1] << 8)
-			+ (msg_read->data[msg_read->readcount + 2] << 16)
-			+ (msg_read->data[msg_read->readcount + 3] << 24);
+		c = (unsigned int)msg_read->data[msg_read->readcount]
+			+ ((unsigned int)msg_read->data[msg_read->readcount + 1] << 8)
+			+ ((unsigned int)msg_read->data[msg_read->readcount + 2] << 16)
+			+ ((unsigned int)msg_read->data[msg_read->readcount + 3] << 24);
 	}
 
 	msg_read->readcount += 4;
